@@ -1,82 +1,60 @@
 {
-  "name": "bfend",
-  "version": "0.3.5",
-  "license": "MIT",
-  "description": "Admin template base on ng-zorro-antd",
-  "keywords": [
-    "ant",
-    "design",
-    "angular",
-    "ui",
-    "framework",
-    "frontend",
-    "admin"
-  ],
-  "homepage": "https://github.com/hypocenter/bfend",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hypocenter/bfend"
-  },
-  "bugs": {
-    "url": "https://github.com/hypocenter/bfend/issues"
-  },
-  "author": "hypo hypomail@gmai.com",
-  "main": "./bfend.js",
-  "typings": "./bfend.d.ts",
+  "name": "bfend-app",
+  "version": "0.4.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "start-with-mock": "node ./mock/server & ng serve",
-    "mock": "node ./mock/server",
-    "site:build": "ng build --prod --build-optimizer",
-    "build": "bash ./build.sh",
-    "release:next": "bash ./build.sh && cd publish && npm publish --access public --tag next",
-    "release": "bash ./build.sh && cd publish && npm publish --access public",
+    "build": "ng build",
+    "build:lib": "sh scripts/build.sh",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "mock": "node ./mock/server",
+    "release": "npm run build:lib && cd dist/bfend && npm publish --access public"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.11",
-    "@angular/common": "^5.2.11",
-    "@angular/compiler": "^5.2.11",
-    "@angular/core": "^5.2.11",
-    "@angular/forms": "^5.2.11",
-    "@angular/http": "^5.2.11",
-    "@angular/platform-browser": "^5.2.11",
-    "@angular/platform-browser-dynamic": "^5.2.11",
-    "@angular/router": "^5.2.11",
-    "@auth0/angular-jwt": "^1.2.0",
-    "angular-web-storage": "^2.0.0",
-    "core-js": "^2.5.7",
-    "glob": "^7.1.2",
-    "moment": "^2.22.2",
-    "ng-zorro-antd": "^0.7.1",
-    "ngx-progressbar": "^2.1.1",
-    "rxjs": "^5.5.11",
-    "simple-line-icons": "^2.4.1",
-    "ya-mock-server": "^0.2.1",
+    "@angular/animations": "^6.0.3",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "bfend": "file:./dist/bfend",
+    "core-js": "^2.5.4",
+    "less": "^2.7.3",
+    "rxjs": "^6.0.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.11",
-    "@angular/language-service": "^5.2.11",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.114",
-    "codelyzer": "^4.4.2",
-    "dotenv": "^5.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular-devkit/build-ng-packagr": "~0.6.8",
+    "@angular/cli": "~6.0.8",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.2.1",
+    "dotenv": "^6.0.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.4",
+    "karma": "~1.7.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.4.3",
-    "karma-jasmine": "^1.1.2",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
+    "ng-packagr": "^3.0.0-rc.2",
+    "protractor": "~5.3.0",
+    "rxjs-tslint": "^0.1.5",
+    "ts-node": "~5.0.1",
+    "tsickle": ">=0.25.5",
+    "tslib": "^1.7.1",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "~2.7.2",
+    "ya-mock-server": "^0.2.1"
   }
 }
