@@ -1,10 +1,12 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { BfHttpService } from 'bfend';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class MeApi {
 
   constructor(private client: HttpClient, private http: BfHttpService) {}
